feat: new hugo server

+++ src-tauri/src/commands/ssh_command.rs
@@ -12,15 +12,16 @@ pub fn kill_server() -> Result<(), InvokeError> {
         &mut channel,
         &format!("ps -ef | grep '{} server' | grep -v grep | awk '{{print$2}}'", &hugo_config.hugo_cmd_path)
     ).into_invoke_err()?;
+    let pid = pid.trim();
     println!("pid: {}", pid);
-    // let s = execute_ssh_command(&mut channel, "id").into_invoke_err()?;
-    // println!("s: {}", s);
-    channel = get_channel_session().into_invoke_err()?;
-    let s = execute_ssh_command(
-        &mut channel,
-        &format!("kill -9 {}", pid)
-    ).into_invoke_err()?;
-    println!("s: {}", s);
+    if !pid.is_empty() {
+        channel = get_channel_session().into_invoke_err()?;
+        let s = execute_ssh_command(
+            &mut channel,
+            &format!("kill -9 {}", pid)
+        ).into_invoke_err()?;
+        println!("s: {}", s);
+    }
 
     Ok(())
 }
